Better approach to detecting locked files. Previous checks were too tight, whereas others were too loose. Implement a mixed approach, which it may be good to generalize in the future.
git-svn-id: https://synchronicity.svn.sourceforge.net/svnroot/synchronicity/trunk@1162 6fc22aba-598d-424d-8c45-4c47160ecdfc

+++ Create Synchronicity/SynchronizeForm.vb
@@ -876,8 +876,19 @@ Public Class SynchronizeForm
             Static GZipCompressor As Compressor = LoadCompressionDll()
             GZipCompressor.CompressFile(SourceFile, DestFile, Sub(Progress As Long) Status.BytesCopied += Progress) ', ByRef ContinueRunning As Boolean) 'ContinueRunning = Not [STOP]
         Else
-            Using TestForAccess As New IO.FileStream(SourceFile, IO.FileMode.Open, IO.FileAccess.Read, IO.FileShare.None) : End Using 'Checks whether the file can be accessed before trying to copy it. This line was added because if the file is only partially locked, CopyFileEx starts copying it, then fails on the way, and deletes the destination.
-            IO.File.Copy(SourceFile, DestFile, True)
+            If IO.File.Exists(DestFile) Then
+                Try
+                    Using TestForAccess As New IO.FileStream(SourceFile, IO.FileMode.Open, IO.FileAccess.Read, IO.FileShare.None) : End Using 'Checks whether the file can be accessed before trying to copy it. This line was added because if the file is only partially locked, CopyFileEx starts copying it, then fails on the way, and deletes the destination.
+                    IO.File.Copy(SourceFile, DestFile, True)
+                Catch Ex As IO.IOException
+                    Dim TempDest As String = DestFile & IO.Path.GetRandomFileName(), DestBack As String = DestFile & IO.Path.GetRandomFileName()
+                    IO.File.Copy(SourceFile, TempDest, False) 'Don't overwrite, in case of a random filename collision.
+                    IO.File.Move(DestFile, DestBack) : IO.File.Move(TempDest, DestFile)
+                    IO.File.Delete(DestBack)
+                End Try
+            Else
+                IO.File.Copy(SourceFile, DestFile)
+            End If
         End If
 
         If Handler.GetSetting(ConfigOptions.TimeOffset, "0") <> "0" Then 'Updating attributes is needed.
